IPRT: socket / address resolving fixes.
svn:sync-xref-src-repo-rev: r75775

+++ src/VBox/Runtime/r3/posix/process-posix.cpp
@@ -1,4 +1,4 @@
-/* $Id: process-posix.cpp 33602 2010-10-29 12:39:54Z knut.osmundsen@oracle.com $ */
+/* $Id: process-posix.cpp 39801 2012-01-18 18:01:11Z knut.osmundsen@oracle.com $ */
 /** @file
  * IPRT - Process, POSIX.
  */
@@ -130,6 +130,9 @@ RTR3DECL(int)   RTProcWaitNoResume(RTPROCESS Process, unsigned fFlags, PRTPROCST
 
 RTR3DECL(int) RTProcTerminate(RTPROCESS Process)
 {
+    if (Process == NIL_RTPROCESS)
+        return VINF_SUCCESS;
+
     if (!kill(Process, SIGKILL))
         return VINF_SUCCESS;
     return RTErrConvertFromErrno(errno);
